+++ src/main/scala/uk/co/gresearch/spark/dgraph/connector/partitioner/AlphaPartitioner.scala
@@ -16,13 +16,15 @@
 
 package uk.co.gresearch.spark.dgraph.connector.partitioner
 
-import uk.co.gresearch.spark.dgraph.connector.{ClusterState, Partition, Schema}
+import uk.co.gresearch.spark.dgraph.connector.{AlphaPartitionerOption, ClusterState, Partition, Schema}
 
 case class AlphaPartitioner(schema: Schema, clusterState: ClusterState, partitionsPerAlpha: Int)
   extends Partitioner with ClusterStateHelper {
 
   if (partitionsPerAlpha <= 0)
     throw new IllegalArgumentException(s"partitionsPerAlpha must be larger than zero: $partitionsPerAlpha")
+
+  override def configOption: String = AlphaPartitionerOption
 
   override def getPartitions: Seq[Partition] = {
     PredicatePartitioner.getPartitions(