+++ src/test/scala/uk/co/gresearch/spark/dgraph/connector/sources/TestEdgeSource.scala
@@ -328,7 +328,7 @@ class TestEdgeSource extends AnyFunSpec with ShuffleExchangeTests
       val withoutPartitioning = () =>
         reader
           .options(Map(
-            PartitionerOption -> s"$UidRangePartitionerOption",
+            PartitionerOption -> UidRangePartitionerOption,
             UidRangePartitionerUidsPerPartOption -> "2",
             MaxLeaseIdEstimatorIdOption -> dgraph.highestUid.toString
           ))
@@ -374,7 +374,7 @@ class TestEdgeSource extends AnyFunSpec with ShuffleExchangeTests
       val withPartitioning = () =>
         reader
           .options(Map(
-            PartitionerOption -> s"$UidRangePartitionerOption",
+            PartitionerOption -> UidRangePartitionerOption,
             UidRangePartitionerUidsPerPartOption -> "2",
             MaxLeaseIdEstimatorIdOption -> dgraph.highestUid.toString
           ))
@@ -410,7 +410,7 @@ class TestEdgeSource extends AnyFunSpec with ShuffleExchangeTests
       val withPartitioning = () =>
         reader
           .options(Map(
-            PartitionerOption -> s"$PredicatePartitionerOption+$UidRangePartitionerOption",
+            PartitionerOption -> PredicateAndUidRangePartitionerOption,
             PredicatePartitionerPredicatesOption -> "1",
             UidRangePartitionerUidsPerPartOption -> "2",
             MaxLeaseIdEstimatorIdOption -> dgraph.highestUid.toString