+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [UNRELEASED] - YYYY-MM-DD
+
+### Changed
+- Node source in wide mode does not allow for any predicate partitioning.
+
 ## [0.8.0] - 2022-01-19
 
 ### Changed